wizard restructure
