--- MoneyStats.WebApp/ClientApp/src/app/components/apply-rules/apply-rules.component.ts
@@ -1,16 +1,0 @@
-import { Component, Input, Output, OnInit } from '@angular/core';
-
-@Component({
-    selector: 'app-apply-rules-component',
-    templateUrl: './apply-rules.component.html',
-    styleUrls: ['./apply-rules.component.scss']
-})
-export class ApplyRulesComponent implements OnInit {
-
-    constructor() {
-    }
-
-    ngOnInit(): void {
-    }
-    
-}
