--- MoneyStats.WebApp/ClientApp/src/app/components/bank-row-to-transaction/bank-row-to-transaction.component.html
@@ -1,1 +1,0 @@
-<p>bank-row-to-transaction works!</p>
