--- MoneyStats.WebApp/ClientApp/src/app/components/bank-row-to-transaction/bank-row-to-transaction.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-bank-row-to-transaction-component',
-  templateUrl: './bank-row-to-transaction.component.html',
-  styleUrls: ['./bank-row-to-transaction.component.scss']
-})
-export class BankRowToTransactionComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
