+++ MoneyStats.WebApp/ClientApp/src/app/components/nav/nav.component.html
@@ -23,7 +23,7 @@
         </a>
         <div class="dropdown-menu" aria-labelledby="navbarDropdown">
           <div title="Read in bank exports">
-            <a class="dropdown-item" [routerLink]='["/read-in"]'>
+            <a class="dropdown-item" [routerLink]='["/input-bank-rows"]'>
               <fa-icon icon="university"></fa-icon>
               <span>Bank exports</span>
             </a>
