--- MoneyStats.WebApp/ClientApp/src/app/components/read-in/read-in.component.scss
@@ -1,132 +1,0 @@
-
-.content {
-    padding: 0 10px;
-}
-
-.tr-table {
-    $cell-height: 30px;
-
-    .tr-row {
-        vertical-align: top;
-        white-space: nowrap;
-
-        .tr-cell:first-child {
-            border-left: 1px solid #ddd;
-        }
-
-        &:last-child .tr-cell {
-            border-bottom: 1px solid #ddd;
-        }
-
-        &.header {
-            font-weight: bold;
-            height: $cell-height;
-        }
-
-        &.excluded .tr-row-content .tr-cell {
-            background-color: #f1f1f1;
-            color: #b6b6b6;
-        }
-
-        &.excluded .tr-row-content .tr-cell:first-child {
-            background-color: transparent;
-            color: initial;
-        }
-
-        &.opened .tr-cell {
-            border-bottom: 1px solid #ddd;
-        }
-
-        .tr-cell {
-            display: inline-block;
-            border-top: 1px solid #ddd;
-            border-right: 1px solid #ddd;
-            padding: 2px 5px;
-            height: $cell-height;
-            white-space: nowrap;
-            overflow: hidden;
-            text-overflow: ellipsis;
-            min-width: 76px;
-        }
-
-        .tr-row-content {
-            height: $cell-height;
-        }
-
-        .tr-details {
-            padding: 10px 0;
-            font-size: 12px;
-
-            .tr-details-menu {
-                .tr-details-menu-item {
-                    display: inline-block;
-                    padding: 5px 8px;
-                    border: 1px solid #ddd;
-                    border-bottom: none;
-                    border-right: none;
-                    cursor: pointer;
-                    background: #ddd;
-                    color: #777;
-
-                    &:hover,
-                    &.active {
-                        color: #111;
-                        background: #fff;
-                    }
-
-                    &.active {
-                        cursor: default;
-                    }
-                }
-
-                .tr-details-menu-item:last-child {
-                    border-right: 1px solid #ddd;
-                }
-            }
-
-            .tr-details-content {
-                padding: 8px 8px 5px 8px;
-                border: 1px solid #ddd;
-                white-space: pre-wrap;
-
-                p {
-                    padding: 0 0 5px 0;
-                    margin: 0;
-                }
-
-                .title-item {
-                    font-weight: bold;
-                }
-
-                .alert-item {
-                    background: #eee;
-                    padding: 2px 4px;
-                    margin-bottom: 5px;
-                }
-
-                .no-alert-item {
-                    color: #777;
-                }
-            }
-        }
-    }
-
-    .tr-btn {
-        display: inline-block;
-        background-color: #444;
-        color: #eee;
-        cursor: pointer;
-        font-size: 15px;
-        height: 25px;
-        line-height: 25px;
-        width: 25px;
-        text-align: center;
-        border-radius: 3px;
-        margin-right: 2px;
-
-        &.excluded {
-            background-color: #666;
-            color: #eee;
-        }
-    }
-}
