--- MoneyStats.WebApp/ClientApp/src/app/components/update-wizard/step-1/choose-file/choose-file.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ChooseFileComponent } from './choose-file.component';
-
-describe('ChooseFileComponent', () => {
-  let component: ChooseFileComponent;
-  let fixture: ComponentFixture<ChooseFileComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ChooseFileComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ChooseFileComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
