--- MoneyStats.WebApp/ClientApp/src/app/components/update-wizard/step-2/read-files/read-files.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ReadFilesComponent } from './read-files.component';
-
-describe('ReadFilesComponent', () => {
-  let component: ReadFilesComponent;
-  let fixture: ComponentFixture<ReadFilesComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ReadFilesComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ReadFilesComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
