--- MoneyStats.WebApp/ClientApp/src/app/components/update-wizard/step-3/compare-db/compare-db.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { CompareDbComponent } from './compare-db.component';
-
-describe('CompareDbComponent', () => {
-  let component: CompareDbComponent;
-  let fixture: ComponentFixture<CompareDbComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ CompareDbComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(CompareDbComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
