+++ MoneyStats.WebApp/ClientApp/src/app/components/wizard-navigation/wizard-navigation.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { RouterService } from 'src/app/services/router-service/router.service';
-import { WizardStep } from "src/app/models/component-models/wizard-step.model";
+import { WizardStep } from '../wizards/wizard-step.model';
 
 
 @Component({
