+++ MoneyStats.WebApp/ClientApp/src/app/components/wizards/create-transactions-wizard/eval-transactions/tag-colorer.ts
