+++ MoneyStats.WebApp/ClientApp/src/app/components/wizards/create-transactions-wizard/eval-transactions/used-generated-transaction.ts
