+++ MoneyStats.WebApp/ClientApp/src/app/components/wizards/input-bankrows-wizard/compare-db/compare-db.component.ts
@@ -5,27 +5,36 @@ import { LoadingScreenService } from 'src/app/services/loading-screen-service/lo
 import { BankRowService } from 'src/app/services/bank-row.service';
 import { BankRow } from 'src/app/models/service-models/bank-row.model';
 import { StaticMessages } from 'src/app/utilities/input-messages.static';
-import { StepAlert } from "src/app/models/component-models/step-alert.model";
-import { CompareDbInput } from 'src/app/pages/update-page/update.page';
-import { Wizard } from 'src/app/components/wizard-navigator/wizard-navigator.component';
-import { WizardStepBase } from '../../step-1/choose-file/choose-file.component';
+import { WizardNavigator } from 'src/app/components/wizards/wizard-navigator/wizard-navigator.component';
 import { Common } from 'src/app/utilities/common.static';
+import { WizardStep } from '../../wizard-step.model';
+
+export class CompareDbInput {
+
+    wizard: WizardNavigator;
+    readBankRowForDbCompare: ReadBankRowForDbCompare[];
+    mapper: ExcelBankRowMapper;
+
+    constructor(wizard: WizardNavigator, readBankRowForDbCompare: ReadBankRowForDbCompare[], mapper: ExcelBankRowMapper) {
+        this.wizard = wizard;
+        this.readBankRowForDbCompare = readBankRowForDbCompare;
+        this.mapper = mapper;
+    }
+}
 
 @Component({
     selector: 'app-compare-db-component',
     templateUrl: './compare-db.component.html',
     styleUrls: ['./compare-db.component.scss']
 })
-export class CompareDbComponent extends WizardStepBase implements OnInit {
+export class CompareDbComponent extends WizardStep implements OnInit {
 
     @Input() $input: CompareDbInput;
-
-    @Output() nextStepChange = new EventEmitter<string>();
-    //@Output() nextStepAlertsChange = new EventEmitter<string[]>();
-
     public get bankRows(): ReadBankRowForDbCompare[] { return this.$input.readBankRowForDbCompare; }
     public get mapper(): ExcelBankRowMapper { return this.$input.mapper; }
-    public get wizard(): Wizard { return this.$input.wizard; }
+    public get wizard(): WizardNavigator { return this.$input.wizard; }
+
+    @Output() nextStepChange = new EventEmitter<string>();
 
     isError: boolean;
 
@@ -49,8 +58,6 @@ export class CompareDbComponent extends WizardStepBase implements OnInit {
             console.log("Compare finished.");
 
             self.checkForAlerts();
-            // self.nextStepChange.emit(self.bankRows);
-            // self.emitNextStepAlerts();
         });
     }
 
@@ -78,29 +85,9 @@ export class CompareDbComponent extends WizardStepBase implements OnInit {
                     fileRow.messages.push(StaticMessages.MATCHING_READ_BANKROW_WITH_DB);
                     fileRow.setToExclude();
                 }
-
             }
         }
     }
-
-    // private emitNextStepAlerts(): void {
-    //     let alerts = [];
-
-    //     // Error messages
-    //     if (this.bankRows.length === 0) {
-    //         alerts.push(new StepAlert("Bankrow count is zero!").setToCriteria());
-    //     }
-    //     if (!this.bankRows.some(x => !x.isExcluded)) {
-    //         alerts.push(new StepAlert("All bankrows are excluded!").setToCriteria());
-    //     }
-
-    //     // Success messages
-    //     if (alerts.length === 0) {
-    //         alerts.push(new StepAlert("By clicking next step, the new BankRows will be saved to database!"));
-    //     }
-
-    //     this.nextStepAlertsChange.emit(alerts);
-    // }
 
     sortBy_bankRows(arr: any[], property: string) {
         return arr.sort((a, b) => this.dateComparer(a.bankRow[property], b.bankRow[property]));
@@ -118,9 +105,6 @@ export class CompareDbComponent extends WizardStepBase implements OnInit {
     click_toggleRowExclusion(row: ReadBankRowForDbCompare): void {
         row.toggleExclusion();
         this.checkForAlerts();
-
-        //this.nextStepChange.emit(this.bankRows);
-        //this.emitNextStepAlerts();
     }
 
     click_switchDetailsMenu(row: ReadBankRowForDbCompare, i: number): void {
@@ -138,7 +122,7 @@ export class CompareDbComponent extends WizardStepBase implements OnInit {
         }
     }
     
-    public next(): void {
+    next(): void {
         if (!this.wizard.isProgressable()) {
             return;
         }
@@ -149,9 +133,10 @@ export class CompareDbComponent extends WizardStepBase implements OnInit {
             self.nextStepChange.emit("Success");
             self.wizard.next();
         });
+    }
 
-        // this.nextStepChange.emit(this.bankRows);
-        // this.wizard.next();
+    previous(): void {
+        this.wizard.previous();
     }
 
     private getIncludedBankRows(toCast: ReadBankRowForDbCompare[]): BankRow[] {
