+++ MoneyStats.WebApp/ClientApp/src/app/components/wizards/wizard-navigator/wizard-navigator.component.html
@@ -12,7 +12,6 @@
         <div class="info">
             <h3>{{infoBoxStep.title}}</h3>
             <p *ngFor="let dsc of infoBoxStep.description">{{dsc}}</p>
-            <p *ngFor="let stepAlert of wizard.currentStep.stepAlerts">{{stepAlert.title}}</p>
         </div>
     </div>
 </div>
