+++ MoneyStats.WebApp/ClientApp/src/app/components/wizards/wizard-navigator/wizard-navigator.component.scss
