+++ MoneyStats.WebApp/ClientApp/src/app/components/wizards/wizard-step.model.ts
@@ -1,0 +1,5 @@
+export abstract class WizardStep {
+    protected abstract checkForAlerts(): void;
+    public abstract previous(): void;
+    public abstract next(): void;
+}
