--- MoneyStats.WebApp/ClientApp/src/app/models/component-models/step-alert-type.enum.ts
@@ -1,6 +1,0 @@
-
-export enum StepAlertType {
-    Criteria,
-    Message,
-    GreenText
-}
