--- MoneyStats.WebApp/ClientApp/src/app/models/component-models/update-results-utilities.model.ts
@@ -1,8 +1,0 @@
-import { ExcelBankRowMapper } from 'src/app/models/component-models/excel-bank-row-mapper';
-
-/// <summary>
-/// A class for utilities used throughout multiple steps.
-/// </summary>
-export class UpdateResultsUtilities {
-    bankMapper: ExcelBankRowMapper;
-}
