--- MoneyStats.WebApp/ClientApp/src/app/pages/apply-rules-page/apply-rules.page.html
@@ -1,1 +1,0 @@
-<app-apply-rules-component></app-apply-rules-component>
