--- MoneyStats.WebApp/ClientApp/src/app/pages/apply-rules-page/apply-rules.page.ts
@@ -1,15 +1,0 @@
-import { Component, Input, Output, OnInit } from '@angular/core';
-
-@Component({
-    selector: 'app-apply-rules-page',
-    templateUrl: './apply-rules.page.html',
-    styleUrls: ['./apply-rules.page.scss']
-})
-export class ApplyRulesPage implements OnInit {
-
-    constructor() {
-    }
-
-    ngOnInit(): void {
-    }
-}
