--- MoneyStats.WebApp/ClientApp/src/app/pages/bank-row-to-transaction-page/bank-row-to-transaction.page.html
@@ -1,1 +1,0 @@
-<p>bank-row-to-transaction works!</p>
