--- MoneyStats.WebApp/ClientApp/src/app/pages/bank-row-to-transaction-page/bank-row-to-transaction.page.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-bank-row-to-transaction-page',
-  templateUrl: './bank-row-to-transaction.page.html',
-  styleUrls: ['./bank-row-to-transaction.page.scss']
-})
-export class BankRowToTransactionPage implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
