--- MoneyStats.WebApp/ClientApp/src/app/pages/read-in-page/read-in.page.html
@@ -1,19 +1,0 @@
-<div class="left-side-menu">
-    <section class="btns">
-        <!-- To top btn -->
-        <div class="btn-item" (click)="click_scrollToTop()" title="Scroll to top">
-            <div class="btn-toggle">
-                <fa-icon icon="chevron-up"></fa-icon>
-            </div>
-        </div>
-        <!-- To bottom btn -->
-        <div class="btn-item" (click)="click_scrollToBottom()" title="Scroll to bottom">
-            <div class="btn-toggle">
-                <fa-icon icon="chevron-down"></fa-icon>
-            </div>
-        </div>
-    </section>
-</div>
-<div class="read-in-page-content right-side-content">
-    <app-read-in-component></app-read-in-component>
-</div>
