--- MoneyStats.WebApp/ClientApp/src/app/pages/read-in-page/read-in.page.ts
@@ -1,23 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-read-in-page',
-  templateUrl: './read-in.page.html',
-  styleUrls: ['./read-in.page.scss']
-})
-export class ReadInPage implements OnInit {
-
-    constructor() { }
-
-    ngOnInit() {
-    }
-
-    click_scrollToBottom() {
-
-    }
-
-    click_scrollToTop() {
-
-    }
-
-}
