--- MoneyStats.WebApp/ClientApp/src/app/pages/update-page/update.page.scss
@@ -1,16 +1,0 @@
-.left-side-menu {
-    .wizard-nav-section {
-        position: absolute;
-        top: 0;
-        left: 0;
-        right: 0;
-        bottom: 0;
-        padding: 15px 0;
-    }
-}
-
-.update-page-content {
-    h2 {
-        font-size: 18px;
-    }
-}
